notification/requirements.txt
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
redis==3.5.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
PyYAML==6.0.1



